+++ backend/canisters/bot_api_gateway/api/src/lifecycle/init.rs
@@ -4,6 +4,7 @@ use types::{BuildVersion, CanisterId};
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct Args {
+    pub user_index_canister_id: CanisterId,
     pub local_user_index_canister_id: CanisterId,
     pub local_group_index_canister_id: CanisterId,
     pub cycles_dispenser_canister_id: CanisterId,