+++ backend/canisters/bot_api_gateway/impl/src/model/mod.rs
@@ -1,1 +1,1 @@
-
+pub mod bot_registry;