+++ backend/canisters/bot_api_gateway/impl/src/updates/mod.rs
@@ -1,2 +1,3 @@
+mod c2c_notify_user_index_events;
 mod call;
 mod wallet_receive;